CLN: Refactor notifier mock in separate files
