+++ cmd/sovereignnode/systemTestDemo/mockNotifier/interface.go
@@ -1,11 +1,11 @@
 package main
 
-// MockServer holds the mock server actions
-type MockServer interface {
+// GRPCServer holds the mock server actions
+type GRPCServer interface {
 	ExtractRandomBridgeTopicsForConfirmation() ([]*ConfirmedBridgeOp, error)
 }
 
-// GRPCServerMock holds the grpc server actions
-type GRPCServerMock interface {
+// GRPCServerConnection holds the grpc server actions
+type GRPCServerConnection interface {
 	Stop()
 }
